Minor changes made while investigating the wave simulation issue
*** These changes do not fix the issue ***

I haven't found a way to fix this issue yet, but these minor
changes helped me rule out some possible causes, so you might
want to commit them or not (your choice).

The bug itself:
Looking at the "simulation.js" file in the wave simulation, the
issue comes from using CLAMP_TO_EDGE on the
initialSpectrumTexture, spectrumTexture, pingTransformTexture
and pongTransformTexture. Changing that to REPEAT or MIRROR
would solve the issue. Since both REPEAT and MIRROR both solve
the issue, it seems like the texture is never accessed
out-of-bounds and the extrapolation mode should have no effect.

Weirdly, in SamplerCore, clamping to 0x10000 instead of clamping
to 0xFFFF solves the issue (visually), but breaks all other
CLAMP_TO_EDGE tests I tried, so it's not an actual solution, but
it seems to trigger something for this particular case. The
clamping operation caused by CLAMP_TO_EDGE should be a NOOP,
but, obviously it's not. I haven't found out why yet.

BUG=18071764

Change-Id: I0ea38ae0417d20349640bbfcf1124af71f311bbf