Update patch set 2
Patch Set 2:

This doesn't affect any tests launched from the IDE, since it immediately detects that the debugger is already attached. For any stand-alone apps using the Debug DLLs we typically always want to attach the debugger. Can't really think of a counter-example, but in the 'worst' case you just tap enter to make the dialog box go away.

Patch-set: 2