Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

Keep in mind that we'll have to implement support for 1D textures at some point. So there's probably going to be a fair bit of repetition that can be eliminated. Anyway, I'm fine with a bit of duplication for now as long as you keep aware of opportunities to merge things later.

More importantly, do we have anything that exercises this code currently?

Patch-set: 1
Label: Code-Review=+1