Update patch set 5
Patch Set 5:

This attempt of hooking up the memory is more a hack to try to get things going as I don't fully understand how image/texture work in SwiftShader. So yes if there's a better coordination between gralloc and SwiftShader that'd be great and that's probably the right way to implement this anyway :).

Patch-set: 5
Label: Code-Review=0