Update patch set 6
Patch Set 6:

Alexis, note that I renamed some of the enums in libGLESv2/Image.cpp to have a slightly better idea of what each of them does. Basically they're a type and then a size, where the size can be either the number of components, named components with their order, or individual component sizes. Let me know if that makes sense or if I made it more confusing. :-)

Patch-set: 6