Update patch set 4
Patch Set 4:

I'd propose a slightly different approach:

There is a just in time debugging feature on Android that is enabled through setprop debug.db.uid <max_pid> . If this property is set and our pid is < max_pid then we can afford to call abort. The framework will spin up a gdbserver automatically and we can attach a debugger.

If it's not set I'd still like to go through with the logging.

I'm testing this approach on the current maps crash.

Patch-set: 4
Label: Code-Review=0