Update patch set 2
Patch Set 2:

In the emulator-like setup this does get me farther. Although there's still hang/crash but doesn't seems to be on the code gen side anymore. I couldn't quite pinpoint where the crash is as it manifests differently every time. However, I assume it's probably not worth spending too much time there if it's not observed in other environments/setups. I can open a bug to track it if it happens again. Thanks!

Patch-set: 2