Update patch set 11
Patch Set 11:

PTAL. I needed to expand the changes to OpenGL/compiler. These changes also implement an Android-specific version of assert() because it was called directly at some points in the code.

All of the remaining asserts are in LLVM. They haven't been going off much, so my guess is that it will be ok to leave them as is.

Patch-set: 11
Label: Code-Review=0