Update patch set 1
Patch Set 1: Code-Review+2

Looks correct to me but the spec is a bit confusing so please double-check:

"GetVertexAttribfv reads and returns the current attribute
values as floating-point values; GetVertexAttribiv reads them as floating-point
values and converts them to integer values; GetVertexAttribIiv reads and returns
them as integers; GetVertexAttribIuiv reads and returns them as unsigned integers.
The results of the query are undefined if the current attribute values are read
using one data type but were specified using a different one."

Note that this means it's typical for attributes that are read back as signed integers, to have been specified as floating-point. Since you're treating everything as a conversion to the requested type, that should work for the demanded behavior and be the most reasonable for the undefined behavior.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+2