Update patch set 1
Patch Set 1:

I realize 2D arrays and 3D textures are very similar, but with mipmaps one is conceptually an array of pyramids while the other is a 3D pyramid of cubes. I'm still trying to wrap my head around whether we can treat those as perfectly equivalent. It's probably just a matter of the order in which the pixels are stored, and accessed...

That said, the AEP also adds support for arrays of cube textures. We're not going to implement that any time soon, but it would be good to know if that could all still be supported with the current approach.

Concretely, please take a closer look at the Mipmap and Texture structs in Sampler.hpp. If those would need fundamental changes to support all of this then we should attempt to plan ahead for it to start off on the right foot. Thanks!

Patch-set: 1
Label: Verified=0