Update patch set 1
Patch Set 1:

Well, I'm aware the mipmaps are different, which was why I added a virtual function for generateMipmaps() for Texture2DArray. I thought that would be sufficient to differentiate between Texture3D and Texture2DArray.

For texture sampling, that's another story. We might be able to leverage what had been done for Texture2D, and simply add a level to it, but that part still need to be done (and isn't part of this cl).

Patch-set: 1