if/else to switch in FrameBuffer
I propose we change the if/else statements
to switch statements in FrameBuffer, before
we add the RGB565. switch is slightly faster
than if/else (although it doesn't really
matter in this case) an I find it more
readable when all conditions check for the
same value. Feel free to reject it if you
don't like the change.

Change-Id: I970a64ffdc18b240239092615f87a59b84c5e364