Printing sizes properly
Sizes weren't getting printed properly
because of their type. In this case,
primarySize, for example, is an
unsigned char, so it gets printed as if
it were a character instead of a number.
Changing it to "primarySize + '0'" also
fails, because then the result of the
operator+ gets interpreted as an int.
To get it to work, I would need either
something like:
static_cast<char>(primarySize + '0')
or
static_cast<int>(primarySize)
I chose the latter.

Change-Id: Ib0718a7a86ef4314f886b653526240aa788fb3df