Replace the .a's with .so's
We currently have two copies of LLVM in the repository and in memory when we
run: one for ES1 and a second copy for ES2. We can save some space by using a
.so to eliminate the second copy. This is also helpful on JBMR1, where we would
have to introduce a thrid copy in libEGL to pacify the build system.

There is also a change to jump through the PLT if the code is compiled with
-fPIC

Change-Id: I3dda8c8dc2908fa4c9dddce193260a8677512f42