Update patch set 1
Patch Set 1:

I'm not sure if I like this better. It grows the code size because several formats use the same components. Of course for R5G6B5 there's no reuse (until we implement A1R5G5B5), but I didn't want to make it inconsistent where the values are being computed.

On the other hand when more formats doing it all before the switch seems a bit bulky. I'm relying on the compiler to optimize that, but for code readability a bit of repetition isn't so bad.

So I'm on the fence about this. Opinions welcome.

Patch-set: 1
Label: Code-Review=0