Update patch set 2
Patch Set 2:

FYI: There's two kinds of masking being implemented here. One is write masking of individual components. So the old color is read and the bits are masked out that will be replaced and reversely the new color keeps the bits that are written out. New masking constants had to be precomputed for 565.

The second kind is for dealing with writing 2x2 pixel quads at a time while some might be outside the triangle's edge. So they are masked in their entirety. I was able to reuse existing masks for 16-bit values.

Patch-set: 2
Label: Code-Review=0