Update patch set 4
Patch Set 4:

The implementation depends on the list of supported pixel formats. These vary by release, and the only way to tell the release at compile time is to use a GCE-specific compile flag.

We also need the same code in other parts of GCE, so we would actually need to keep two copies of the code updated if we put this in SS.

Patch-set: 4
Label: Code-Review=0