Update patch set 3
Patch Set 3:

This is a clean version of the patch containing two tools for SwiftConfig, a thread-analysis tool and a resource contention tool. I am not sure if you want to land this patch (for future use-cases with MTaaS). If you do, let me know and I will also implement it for ES2.

Patch-set: 3
Label: Code-Review=0