Update patch set 4
Patch Set 4: Code-Review-1

(6 comments)

Instead of deriving the Pixel* Registers from the Rasterizer Registers and requiring a virtual method to allocate them, I think we can just keep them separate. It's nothing more than a structure of Reactor variables so they are passed around more easily as a single argument. So the Rasterizer "Registers" can just be "PolygonData" or something like that, and be passed around as "p" instead of "r". In some places we'll need both "p" and "r" to be passed as a parameter, but that still seems a bit cleaner.

Patch-set: 4
Label: Code-Review=-1