ASTC decoder
This cl adds a partial checkout of the astc-encoder github repo
and uses that code to perform ASTC decoding. Currently, there's
no need for ASTC encoding, so that part of the repo was not
included. The sRGB conversion is not done in the decoder, but
rather done in place, after the decoding is done, directly inside Surface::decodeASTC(). All related dEQP tests pass.

Change-Id: I5f29ec2aa57b04871e839e27a858d5df463f26bd