Update patch set 1
Patch Set 1: Code-Review+2

We have to be a little careful here that TSymbolTableLevel::nextUniqueId() still produces unique IDs after parsing is done. I checked and noticed that it uses a global that is never reset, so we're fine.

TSymbolTableLevel is a bit of an odd class to be handing out globally unique IDs, but that's something we can address later.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+2