Added new parameters to Sampler's state
The new parameters are: minLOD, maxLOD, compFunc, compMode,
baseLevel, maxLevel, swizzleR, swizzleG, swizzleB, swizzleA.

These parameters are sent from the context, either originating
from the Texture object or the Sampler object, if there is one,
and are propagated all the way down to the Sampler's state.

Since it was trivial to do so, an implementation for the
swizzleR, swizzleG, swizzleB, swizzleA parameters was done in
SamplerCore. All related dEQP tests pass.

Change-Id: I45405a0d241d0e70a91a3c56357c7bc04c7dc75a