Update to number of allowed shader inputs/outputs for OpenGL ES 3
Because of the 3 reserved uniforms used for gl_DepthRange, the
number of uniforms allowed in fragment and vertex shaders was
too small by 3, so their allowed number was increased by 3.

As for attributes and varyings, their numbers needed to be
increased in order to pass all OpenGL ES 3.0 dEQP tests.

Change-Id: I6527334e1503c08281303915654087cbac8089a0