Update patch set 3
Patch Set 3:

I'm sure. This affects only the debug build, and we ship only release at the moment.

It's impact on debugging is large because we can get a full stack backtrace when things go wrong.

It's impact on performance is rather modest for multiple reasons:

o We're running only on 1-2 core instances
o SwiftShader uses a different set of threads in each process
o There's already a lot of logging in the debug build, and that seems to be the performance-limiting factor

This is part of my work to standardize the builds and make debugging easier. This is rather important given the number of groups that are taking interest in the GCE work.

The next step is to place both the debug and release SwiftShader on the image and allow the debug version to be activated by a metadata change and reboot.

Patch-set: 3
Label: Code-Review=0