New Short->Int and UShort->Int constructors
Added 2 new constructors for the Int4 class in Nucleus,
which are for conversion to Int4 from Short4 or UShort4.
Some of the required code was extracted from Float4
constructors, which were cleaned up to remove the
duplicate code.

All 4 constructors (2 new Int4 and 2 modified Float4)
behaved properly in dEQP tests.

Change-Id: Ib400ea27a867d31901f4a9ca08bf6093f186b277