Update patch set 2
Patch Set 2:

Weird, the spec is pretty clear:
"If the value of UNPACK_ROW_LENGTH is zero, then the number of groups in a row is width; otherwise the number of groups is UNPACK_ROW_LENGTH."
If UNPACK_ROW_LENGTH was unused, this change should be a noop.
If it is used as a number of bytes instead of a number of pixels, then it is used incorrectly.

Patch-set: 2