Update patch set 4
Patch Set 4:

Alexis: Indeed it's a bit annoying to have an extra dependency besides Git for Windows, but Bison 2.4.1 dates back to 2008. It looks like the binary that is bundled with Git for Windows comes from http://gnuwin32.sourceforge.net/packages/bison.htm. Nobody is maintaining it, while the cygwin version is well maintained. On Linux we'd also be using the more recent version.

I'll ping the Git for Windows developers to see if they can update Bison so we can avoid needing cygwin...

Patch-set: 4
Label: Code-Review=0