Update patch set 2
Patch Set 2:

Unless the configs are differentiated by anything else (e.g. the native visual ID), these new configs are actually ignored because only the sizes of the components are stored as attributes, and the Display stores a set that removes any duplicates. Thus this patch should not cause any difference in behavior (until we differentiate these configs).

Patch-set: 2