Update patch set 2
Patch Set 2:

Ok, so what would the desired behavior be then? I assume we want to keep the static checks (using r.out), which will represent the most common cases, to avoid putting unnecessary code in the shader. So is the desired behavior that if we have any dynamically chosen output in the code, we write to all outputs? (basically, you want me to eliminate the "If((r.dynOut..." and associated code?)

Patch-set: 2