Update patch set 1
Patch Set 1:

This can indeed be a partial fix, because the Y-plane was using the stride (pitchP) while the U en V-plane were basing the stride on the width. That said, they *should* be the same, and if they're not I blieve that would be a gralloc bug. I still think ag/816285 is also part of the solution here. Have you tried that together with this change?

Patch-set: 1