Update patch set 2
Patch Set 2:

Removed the isRenderbuffer() stuff. According to the 3.0.4 spec, section 4.4.2 (about GenRenderbuffers):
"These names are marked as used, for the purposes of GenRenderbuffers only, but they do not acquire renderbuffer state until they are first bound."
So, unlike Sampler objects, calling the Gen... function does not allow the application to interpret this as an object until it is bound.

Patch-set: 2