Update patch set 3
Patch Set 3:

I refactored it a bit to not require the GCE_PLATFORM_SDK_VERSION, but instead define ANDROID_PLATFORM_SDK_VERSION as part of one of SwiftShader's .mk's.

I also replaced the macros with (inline) functions, and I'm calling native_window_dequeue_buffer_and_wait to not have to deal with waiting on the fence (for SDK > 16).

I got rid of the android::NO_ERROR and the header that was needed for that, since these functions return 0 on no error.

Patch-set: 3