Update patch set 2
Patch Set 2:

It can be done. This would mean that x and y would be expressed in a way, and w would be expressed in a different way though, but that's probably expected for texture arrays. We'd still need to clamp the value between 0 and (depth - 1), this can be done with the address function.

Patch-set: 2