Update patch set 1
Patch Set 1:

I was planning to use the magFilter to implement this portion of the OpenGL spec (section 8.15, Texture Magnification):
"... make c depend on the combination of minification and magnification modes as follows : if the magnification filter is given by LINEAR and the minification filter is given by NEAREST_MIPMAP_NEAREST or NEAREST_MIPMAP_LINEAR, then c = 0.5. This is done to ensure that a minified texture does not appear sharper than a magnified texture. Otherwise c = 0."

Patch-set: 1
Label: Code-Review=0