Update patch set 5
Patch Set 5: Code-Review+2 Verified+1

This was necessary to fix support for the Android Emulator. It sets the size of the viewport in pixels (as expected), but the width and height of the NSView or CALayer are in "points". This patch translates them to pixels.

Patch-set: 5
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+2
Label: Verified=+1