Update patch set 2
Patch Set 2: Code-Review-1

While this looks fine in and of itself, I suspect that this code will change a lot to fully support uniform buffers. Also note that VERTEX_UNIFORM_VECTORS is the total number of uniforms when it's a dedicated single array, but when we support uniform buffers this becomes much less meaningful.

At the processing routines level we ultimately only care about the maximum number of uniforms to prevent reading out of bounds. MAX_UNIFORM_BLOCK_SIZE appears to be that limit, in bytes. I guess a constant defining the max in float4 vectors would make sense for us (same limit for fragment or vertex).

The spec seems to specify that glUniform() calls use a default uniform buffer that is part of the Program state. Its size is not determined by MAX_UNIFORM_BLOCK_SIZE but by the legacy MAX_VERTEX_UNIFORM_VECTORS. I think it would make the most sense to make it the same size. The processing routines should be oblivious to the difference between the default uniform buffer and named uniforms buffers.

Named uniform buffers have an index starting at zero. That's annoying since we can't use 0 as the default buffer binding, but we can just place it at the end instead (index MAX_UNIFORM_BUFFER_BINDINGS).

Patch-set: 2