Update patch set 3
Patch Set 3: Verified+1

Not quite. It removes all uses of argument() used after a call to emit(), but it doesn't remove uses of argument() combined with lvalue() or used for the ".rel" portion or the Parameter. That leaves about 5 calls to be eliminated, but these are less trivial than the ones eliminated here.

Patch-set: 3
Label: Verified=+1