Update patch set 5
Patch Set 5:

Why do you need this to be active on all builds? This will only help indicate that some crash in the wild is a SwiftShader issue. We'd still need to reproduce it and run it locally to find the real cause and fix it. And for that we can use a Debug build.

My last change makes it a GCE-only feature. I doubt the emulator has a need for this. But I'd still like to understand why you'd like this on non-Debug builds for GCE. I'm just worried about death by a thousand cuts (this apparently invokes the kernel, and the tags are leaked).

Thanks for explaining why it's useful to tag each block with its start address. I just hope the kernel can deal with that many different tags. For a Debug build I wouldn't be concerned about it.

Patch-set: 5
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=0