2D array filtering crash fix
2D array textures aren't supposed to filter between layers and the
computation used in the filtering code is meant to work with 3D
textures, not 2D array textures. There's surely a more efficient
solution than this one for 2D array textures, since the proposed
solution essentially computes the filtering twice (with 8 texels,
instead of only using 4) by using the same path as 3D textures,
but at least this fixes the filtering crashes for now.

Change-Id: Iea8fcfe67cce7ae70e39081af07e67fa59b268ff