Fixed FenceSync object ref count
FenceSync objects are ref counted objects, so the are deleted
using the release() method. This means that they need to be
refed at creation (since we only store the pointer inside the
map), otherwise we'll hit the assert at deletion.

FenceSync objects are currently only ever refed once (at creation)
and unrefed once (at deletion), but are not fully implemented yet.

Change-Id: Ie693ace8895b6140ae96f9325d5c48951c7e9730