Remove FORMAT_A1R5G5B5 from EGL configs.
This is not a natively supported pixel format, thus causing a performance
penalty for conversion. It is not supported by Nexus 4 or 5. Android has
deprecated both 4444 and 5551 as HAL formats. Also, UNSIGNED_SHORT_5_5_5_1
has the alpha component in the least significant bit, whereas A1R5G5B5 has
alpha as the most significant bit.

Bug 27690616

Change-Id: I350f0e780cd87b8a626e2c8914a65c61dbf63258