Shader input/output constant cleanup
The number of shader inputs and outputs supported were represented
by constants defined all over the place, sometimes hardcoded values,
to represent different values that were often linked together. Now,
if possible, constants are defined at the end of enums that represent
the list of registers present in a certain register array used for
inputs or outputs in vertex or fragment shaders. Values that are
linked together logically are also represented as an expression of
values they depend on.

Also to note, register array sizes are generally increased by 8 in
this cl. The reason for this is that the number of registers used
in many of the input or output register arrays were based on the
assumption that there were 8 texture units present. Now that the
sizes of register arrays are linked directly to the values of
TEXTURE_IMAGE_UNITS and VERTEX_TEXTURE_IMAGE_UNITS, which are both 16,
the number of texture related registers in register arrays has doubled.

Change-Id: I68963fbf71a269574012786e04c1266beafd90e0