Shader input/output constant cleanup
The number of shader inputs and outputs supported were represented
by constants defined all over the place, sometimes hardcoded values,
to represent different values that were often linked together. Now,
if possible, constants are defined at the end of enums that represent
the list of registers present in a certain register array used for
inputs or outputs in vertex or fragment shaders. Values that are
linked together logically are also represented as an expression of
values they depend on.

No values should have been modified in this cl.

Change-Id: I68963fbf71a269574012786e04c1266beafd90e0