Adding task data for transform feedback
Transform feedback requires a bit of extra data to write properly
into the transform feedback buffers. First it needs to know where
in the buffer to read, which can be derived from the vertex number
provided in vertexStart. Also, since SwiftShader always processes 3
vertices per primitive, regardless of the primitive type, transform
feedback needs to know which vertices to use, and that information
can be derived from verticesPerPrimitive.

Change-Id: I820d99949d7b2955794cc143ffb178e76dd418d7