Update patch set 1
Patch Set 1:

I agree it's a lot of non-trivial code, but it's not an easy problem to get right. And I don't want to settle for a fix that merely passes the test, because the test is the lowest common denominator between shipped drivers, or less, and not necessarily what the spec says. Where possible and reasonable, we should aim to be a reference implementation. That way we also avoid having to come back to this issue when the test gets tighter, or worse, when an app expects it but we first have to track down that this is the issue.

Note that the min/max values are static const, so they're only computed once.

The reason why the second loop only iterates once is obvious when considering that the next 23-bit mantissa float near the magnitude of a 32-bit integer is always a larger integer value. So I could eliminate it. Which leads me to consider replacing the whole thing with a precomputed constant...

Patch-set: 1