Fix checking all mipmap levels for padding.
When textures have padding after a row of pixels, we can't use the sampler's
fast path which uses shift operations. We have to treat it as an NPOT texture.
Previously only mipmap level 0 was checked for padding, but when a texture is
also a render target we allocate 2x2 pixel blocks so a 1x1 mipmap has stride 2.

Change-Id: I7421fddbe5ed0d330d881f09c2161d9c42348600