Switch implementation
Implemented switch/case for glsl in OpenGL ES 3.0.
For simplicity, it is implemented as a single iteration loop so that
break statements work properly like so:

condition = true
for(;condition;) // infinite loop
  if(...) // 1st case
  ...
  else if(...) // other cases
  ...
  else // default case
  ...
  condition = false // unconditionally exit loop
end for

All related dEQP tests pass, except 4 tests where vertex shaders
contain a switch or a loop within another switch. These 4 failures
have only about 5% of bad pixel and seem to be related to a precision
issue, since the equivalent tests inside the fragment shader pass.

KNOWN ISSUE: If a switch is within a loop and one of the cases
             contains a "continue" statement, this will not be
             handled correctly at the moment. There are no dEQP
             tests for this at the moment, AFAIK.

Change-Id: I3ba34ab06a759d07e8520f6a87d75036a5cdaef5