Update patch set 1
Patch Set 1:

All resource memory is allocated using allocateZero(). So it won't access uninitialized data as long as we're within the slice bounds. As Meng-Lin noted, it's undefined behavior so even crashing would in theory be allowed (but the least expected).

That said, the GL_KHR_robust_buffer_access_behavior extension requires returning zero on out-of-bounds fetches. We don't support that extension, but that makes Chrome modify the shaders to perform clamping for WebGL, so we may be able to do it more efficiently ourselves.

I'll leave it up to Alexis to make the judgement call whether it's best to move ahead with other dEQP-GLES3 tests or deal with issues like these while we encounter them.

Patch-set: 1
Label: Verified=0