Update patch set 1
Patch Set 1:

> To date, a reliable way to check format availability is to create a
 > texture with this format and if it generates an OpenGL error, it's
 > not supported... We do that too. We do the same for renderable.
 > 
 > Would it be possible to check each version of OpenGL/ES and
 > extensions to figure out supported formats and renderable?
 > Probably... I can picture that game developers don't like to reach
 > 20 specs. The OpenGL error approach is also very effective against
 > drivers bugs.
 > 
 > My understanding is that this patch is designed to embrace how
 > actual engines works for larger compatibility.

Hi Christophe, great to see you have an interest in SwiftShader!

For what it's worth, I haven't seen popular engines hit this renderability assert much at all, and on the rare occasions when they did it was an indication of a bug in SwiftShader. Hence I'd like to keep it in place.

Indeed to avoid hitting it the engine would have to not check renderability of formats that can't be rendered to (like compressed texture formats), and to be sure you'd need to read all the extension specs. But that doesn't seem like an unreasonable expectation to me. Only the engine developers would have to do so, and shielding the application developers from these low-level details is their responsibility anyway.

Patch-set: 1