Update patch set 2
Patch Set 2: Code-Review-1

(13 comments)

A good start in general, but please try to simplify your loops in order to minimize the checks with the primitive restart indices. It should be possible to write most of these with 1 check per loop while keeping a rolling window for the values you need.

Patch-set: 2
Label: Code-Review=-1