Changed SwiftShader output
To avoid having a library name mismatch between the LIBRARY lines of
the .def files and the output from the GN files, the output names
were restored to libEGL and libGLESv2, but the output directories
were changed to avoid clashing with the ANGLE libraries.

Also removed the warning silencing that's no longer required.

Change-Id: Ifa13d302ffa19acc04e46e4c772fa407a1c818bc