Update patch set 1
Patch Set 1:

> If one is to go by what specs says, glTexImage2D specs is clear
 > about what error will be generated in case invalid enum is passed
 > for texture format:
 > 
 > http://docs.gl/es2/glTexImage2D
 > 
 > "GL_INVALID_ENUM is generated if format or type is not an accepted
 > value."

Hi Branimir. Sorry for the late reply. I was on paternity leave for a month.

Our glTexImage2D implementation does generate GL_INVALID_ENUM when an invalid enum is used (and just that) as far as I can tell. glCheckFramebufferStatus​ is also spec compliant, in a Release build. Whether or not there can be a verbose error in Debug builds is a matter of opinion and not defined in any spec.

Removing the assert wholesale is not a great option to me, but I'm still very open to updating these switch statements with the formats for which you're hitting the assert! If you could provide me with a list of these formats, I'd be happy to write the patch. Thanks.

Patch-set: 1