Fixed a uniform buffer regression
The following cl had broken all fragment uniform buffers:
https://swiftshader-review.googlesource.com/6950

Since the null pointer check was only made on the vertex
uniform buffer, it was preventing the fragment uniform
buffer from being set, which was causing a crash when this
happened. Also, it is safer to call
VertexProcessor::setUniformBuffer and
PixelProcessor::setUniformBuffer unconditionally in order
to set the null pointers correctly when no buffers are set.

Change-Id: I275b0bca6173ee3df4e0e369926ed0a94500e1ba