Add type information to pointer arguments.
Unlike LLVM, Subzero does not store the type of value a pointer
references as part of the pointer's type information. So add a type
argument to createGEP(), even though at the moment we only support
pointer arithmetic on Pointer<Byte>. Also add the type to
createStore(), to enable smaller types to be emulated by larger ones.

Bug swiftshader:11

Change-Id: I08173dd0ba07362d2b27baff4a8fba0ecce093d2