Added support for borders around Surface object buffers
A border is expressed in pixels and will be added on
all edges, resulting in a Surface of dimensions
(w + 2 * border) X (h + 2 * border).

The surface will still expose dimensions of w X h
through the API and will point to the same pixel when
the lock() function is called, so pixel (0,0) will be
returned at the correct offset if a border is present.

As a result, functions within the Surface class that
were using the buffer member directly have been changed
to use the lock() and unlock() functions, so that the
memory always points at the correct pixel.

Cube textures will now have a border around each face,
for use later for linear interpolation between faces.

Change-Id: I53228aa9adad3221d57580c62736fe92671b1548