Add support for borders around textures.
Borders are required to support seamless cubemap sampling. Subsequent
patches will fill the borders with pixels from adjacent cube faces.

The border is expressed in pixels and is added on all edges,
resulting in an image of dimensions
(width + 2 * border) x (height + 2 * border).
The surface still exposes dimensions of width x height through the API
and points to the same pixel when locked.

Change-Id: I53228aa9adad3221d57580c62736fe92671b1548