New utility function to update edges of cube textures
The cube texture edges will be used for linear
interpolation, in order to produce seamless edges.

TextureCubeMap::updateBorders() will update all 4 edges
of each face, so 24 edges will be updated in order to
allow proper linear interpolation.

Image::CopyCubeEdge() will figure out the proper way of
copying the edge based on which edges are being copied
to and from.

Change-Id: Idd17c72c6aaf7dcc65188b065ac8ba179b58cc37