New utility function to update edges of cube textures
The cube texture edges will be used for linear
interpolation, in order to produce seamless edges.

TextureCubeMap::updateBorders() will update all 4 edges
of each face, so 24 edges will be updated in order to
allow proper linear interpolation.

Image::CopyCubeEdge() will figure out the proper way of
copying the edge based on which edges are being copied
to and from.

Change-Id: Ic85aa2b72f8bb4863d70d142a6b7082b0f7d296c