Perform rasterization in floating-point.
Using 12.4 fixed-point coordinates caused issues with very large
triangles due to integer overflow. The delta between the coordinates
requires 17 bits to represent, which when multiplied by a 16-bit value
overflows a 32-bit integer. We can use single-precision floating-point
instead without precision issues when multiplying the slope by the y
delta for every scanline. This keeps the sub-pixel error in the 1%
order of magnitude for a 4096x4096 viewport with 16x16 subpixels.

Bug b/34078120

Change-Id: Ifff8e6085866fd72a96471c4539e675e615b695c