Update patch set 2
Patch Set 2:

Broke Linux build. The function definitions for the typeinfo() functions need to be available at link time. So for example egl::Image's key method needs to be defined in a source file compiled by libEGL. It does not compile Image.cpp, so moving it there broke the linking.

Patch-set: 2