Fixing the Linux build without breaking the Mac build
This change adds 2 platform specific files for Linux, which
are used to define the typeinfo() functions used in some
classes in order to satisfy the ubsan compiler on Linux.
Unfortunately, adding those definitions on all platforms
breaks the build on Mac. Those aren't necessary on Windows,
so the Linux specific files are enough to make the build
work properly on all platforms.

Change-Id: If731ab2e5df47731695412da04f9b2db576ea140