Fix not canceling an already queued graphics buffer.
cancelBuffer() returns a graphics buffer to the buffer queue but
doesn't pass it on to the consumer side. There's currently no reason
for us to call it when the EGL surface gets destroyed or resized, since
we only briefly dequeue the buffer to blit the already rendered result
and immediately queue it again. Even if the queuing fails, we shouldn't
try to cancel the buffer because that's essentially the same as a
queueing but without valid content. It would fail with the same error,
or worse, transfer ownership twice and mess up the reference counting.

Bug b/34981145

Change-Id: Ia37a711db5ebd61bcde14cf4bc45097d4e2d33e8