Work around unsuppressable GCC error.
GCC warns against bitfields not fitting the entire range of an enum
with a fixed underlying type of unsigned int, which gets promoted to
an error with -Werror and cannot be suppressed. However, GCC already
defaults to using unsigned int as the underlying type of an unscoped
enum without a fixed underlying type. So we can just omit it.

Bug chromium:697528

Change-Id: I8db8fef09e3397d4ec5ebe9f7c3c5e4ae9c0edff