Clarify when we set the EGL_FRAMEBUFFER_TARGET_ANDROID attribute.
The EGL_ANDROID_framebuffer_target extension description is vague about
which configs are supposed to have the EGL_FRAMEBUFFER_TARGET_ANDROID
attribute set. The HWComposer isn't supposed to do format conversions,
nor should we have to blit or convert on eglSwapBuffers, so the flag
is only set when the config's format matches the framebuffer format.

Bug b/21804123

Change-Id: Ie348750fbf1b2f096cdf36358d5c90b6f42e30c1