Normalized 8 bit filtering precision fix
For 8 bit values computed into shorts internally, rounding errors can
cause a value of 0xFF, for example, to vary between 0xFF00 and 0xFFFF
within the short representation. That value is then normalized by a
factor of 1.0f / 0xFFFF, which can cause a value of 0xFF to not be
normalized to exactly 1.0f. To fix this, a 0xFF80 mask was added in
order to remove the extra bits that were causing the normalization to
return different results depending on rounding errors and the
normalization factors were adjusted accordingly. The reason for the
extra 9th bit in the mask is for the rounding which happens later on
in the pipeline to still work properly.

Bug swiftshader:39

Change-Id: Iefd1e4275bc12f2c9c329b65a103cbe3d30adb53