Fix locking a destroyed mutex.
Surface should not lock the resource of a parent texture at destruction, because
it can already have been destroyed. For example when a texture's image was bound
as a render target, and the texture is deleted by the app, then the image holds
the last reference to the texture. When the render target image gets deleted, it
first releases its parent texture, and then the underlying surface gets
destroyed.

This patch effectively reverts commit 8a4f6fc48ae52c79f6d3478d0bd04285700e3fcd.
It prevented crashes due to textures getting deleted while threads were still
rendering to it. That issue was fixed more robustly by reference counting the
images. So we no longer need the synchronization for this scenario.

Bug chromium:716803

Change-Id: Ifc3685dcf9e25e8419000af65d4bb7407f26bbcb