Make FrameBuffer class abstract.
Sanitizer tools desire having the definitions of all the class's
methods, even when they're not called in the current linkage unit. To
work around this, make the class abstract so the tools know it can't
be instantiated directly. In the case of sw::FrameBuffer, this can be
achieved by making the destructor pure virtual. Note that it still has a
non-empty definition, since all non-default destructors need a defintion,
and it will get called as part of the destructor chain.

Bug swiftshader:31

Change-Id: I7601b1a725c513ff484cd34a8965636b7f21513c