Add libc++ dependency for libEGL and libGLESv2
This CL adds a temporary dependency on libc++ from libEGL
and libGLESv2.  This is needed to switch Chrome on Linux
to using libc++.  A long-term solution is to depend on
exe_and_shlib_deps.  However, that also pulls in sanitizer
dependencies, which these libraries currently have trouble
compiling with [1].  Once that is fixed, the libc++
dependency can be replaced with exe_and_shlib_deps.

[1] https://bugs.chromium.org/p/swiftshader/issues/detail?id=31

BUG=chromium:723069
R=capn@google.com

Change-Id: I64f0dde61c2b0c263b6a578556c45bdeac10631e