Linear filtering fix
When copying a texture 1:1 with linear filtering enabled, the
texture coordinates fall right at the center of each pixel. This
is causing an edge case where offsetting by half a pixel in each
direction (using SamplerCore::offsetSample()) can overflow into
the previous or next pixel erroneously. 0x8000 is just over half
of 0xFFFF, while 0x7999 is just under half of 0xFFFF, which
prevents the involuntary overflow when sampling from the center
of the pixel.

Change-Id: Ic5175f11f3106943f56113bf5f0dc694e822f59f