textureoffset refactoring
textureoffset was being computed once per component within computeIndices(),
which is called once or twice per sampleTexel() call, which is called many
times when filtering is enabled. Moved it to address(), where it is computed
only once per pixel.

This cl also moves the address function closer to having a floating point
implementation.

Change-Id: I2781c29efdfbf607362fb94f84a05c1b447529ce