Update patch set 2
Patch Set 2: Code-Review-1

Thanks for writing this. As you noted, it changes the order in which the operations are specified. That doesn't affect LLVM because it uses a DAG to represented SSA form and it gets linearized the same way, but it may affect Subzero and alter performance.

A potentially bigger issue is that the number of components gets propagated, and a vector resulting from a sampling operation can undergo more arithmetic operations involving other components. This is currently not the case since only SamplerCore uses these new Vector4f overloaded operators, but it's not obvious that future refactorings in ShaderCore shouldn't use these without resetting the component count.

So this would need some analysis and redesign, but it may not be worth it.

Patch-set: 2
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=-1