Update patch set 2
Patch Set 2:

This is debatable. The default value only applies to RGB components and not to the Alpha component. Could be named "defaultNonAlphaComponentValue", but that's a bit long winded. "defaultColorValue" was chosen as a shorter version of "defaultNonAlphaComponentValue". Not an ideal name.

Feel free to write a cl with the name change if this really bothers you.

Patch-set: 2
Reviewer: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>