Remove sub-vector load/store address hack.
The Optimizer used to assume that loads and stores using the same
address are of the same tpe. This is false for emulated sub-vector
load/store, and thus we tricked it into thinking the addresses are
different by performing some redundant arithmetic. This is no longer
necessary now that the Optimizer checks the types being loaded/stored.

Bug swiftshader:48

Change-Id: Ic212d67fefd9e8dd902ff576b69458208c3c379d