Master CL for Android Emulator 27.2.x
To be abandoned after we switch to glReadPixels and have merged
safe multithreading in its final form.

------------------------------------------------------------
glGetTexImage basic support for GLES2

This patch adds basic support for glGetTexImage in GLES2. It only
supports 2D/Cube textures and a limited set of pixel types.

Rebase on master:

bug: 72158514

Some degree of format conversion is supported / needed for proper
functioning on e.g. API 27 with a GL_RGBA16F app drawer background
texture that is uploaded from a RGBA32F buffer (hence, it would
naturally be downloaded as RGBA32F as well).

+ Bind transform feedback 0 on deletion of current transform feedback
binding. Fixes dEQP-GLES3.functional.lifetime.delete_bound.transform_feedback

------------------------------------------------------------

Prevent crashing when no current program is set.

This fixes a regression caused by
https://swiftshader-review.googlesource.com/14489

Context::applyVertexBuffer() expects a valid program, so we need to
check for it sooner. Not having a current program is not an error and
makes the draw call a no-op, but
dEQP-GLES3.functional.negative_api.vertex_array.* still expects
validation to happen before leaving the function. In particular,
the framebuffer completeness check has to be performed. Note that
ConvertPrimitiveType() should never return an error because we already
check the primitive type in the entry function, sampler validation
can only be done when we have a valid program, and likewise
applyVertexBuffer() requires a program, so these validations can still
happen after the early-out.

Reviewed-on: https://swiftshader-review.googlesource.com/17228
Tested-by: Nicolas Capens <nicolascapens@google.com>
Reviewed-by: Lingfeng Yang <lfy@google.com>
Reviewed-by: Nicolas Capens <nicolascapens@google.com>

------------------------------------------------------------

WIP-thread-safe

Note: rebased on ToT

bug: 74108717

Fixes dEQP-EGL.functional.sharing.gles2.multithread.random failures.
The EGL spec states that "EGL and its client APIs must be threadsafe"
but also that "Client API commands are not guaranteed to be atomic".
Still, dEQP expects that e.g. during the compilation of a shader a
second thread is allowed to delete said shader and we shouldn't even
produce an error on the compile.

Change-Id: I9417e4bbcbded467d4b0a087c80dc81787739713