glGetTexImage basic support for GLES2
This patch adds basic support for glGetTexImage in GLES2. It only
supports 2D/Cube textures and a limited set of pixel types.

Rebase on master:

bug: 72158514

Some degree of format conversion is supported / needed for proper
functioning on e.g. API 27 with a GL_RGBA16F app drawer background
texture that is uploaded from a RGBA32F buffer (hence, it would
naturally be downloaded as RGBA32F as well).

Change-Id: I9417e4bbcbded467d4b0a087c80dc81787739713