Refactor returning a value after generating an error.
Having the return value as part of the error() function was slightly
confusing because it's not always entirely clear which of the arguments
is the error code and which is the return value, and required realizing
that it's a templatized function. In addition, subsequent patches will
add additional arguments to error(). Separating the return value using
the comma operator requires no knowledge of error()'s workings.

Change-Id: I0439d208dfe4563202f13a0e92d15188a84bae9c