Construct RValues from matching LValues types.
Previously any RValue type could be constructed from any LValue type,
or literally anything else that could be substituted and compile. While
we were only implicitly using this to change the signedness of types,
which is safe, it's cleaner to just use an explicit bitcast, and
provide type safety to prevent real issues.

Bug swiftshader:48

Change-Id: Idd1c7f5b9a355dd42220ffb0333d2b88ca467677