Allow Chromium to create an OpenGL ES3 context
Removing the "STRICT_CONFORMANCE" restriction from Chromium.
This allows OpenGL ES3 context creation when using the flag
"--use-gl=swiftshader", which is the case when running layout
tests.
This does not allow WebGL 2 content to be used when using the
flag "--disable-gpu". This will require extra validation on the
Chromium side.

Change-Id: Ic4770c1acbc454a954e1ce813ebed9ed3aee5cd2