WIP-thread-safe
Note: rebased on ToT

bug: 74108717

Fixes dEQP-EGL.functional.sharing.gles2.multithread.random failures.
The EGL spec states that "EGL and its client APIs must be threadsafe"
but also that "Client API commands are not guaranteed to be atomic".
Still, dEQP expects that e.g. during the compilation of a shader a
second thread is allowed to delete said shader and we shouldn't even
produce an error on the compile.

Change-Id: I9e11d45662471163f64fb80d2099ec77e317e861