Build SwiftShader with BOARD_VNDK_VERSION=current
This is using cutils/log.h which is deprecated in O (in favor of
log/log.h) and this is causing build error when building with
BOARD_VNDK_VERSION=current set. However, since SwiftShader should be
able to be built with older versions Android some of which don't have
log/log.h, we can't simply change cutils/log.h to log/log.h.

Instead, liblog_headers is added to the header lib dependency (only for
O and beyond) so that log/log.h can be correctly included via
cutils/log.h

Bug b/63135587

Change-Id: I763250e7410025b1dcd7ae210693e3d6bffdb6f1