Fixed TSAN errors regarding shared memory accesses between threads
TSAN detected many data race errors in the SwiftShader Renderer
class, so in order to fix these, a new AtomicInt class was added
which is a basically a wrapper class for std::atomic<int> and which
only exposes the portion of the API required by SwiftShader.

Since std::atomic isn't available on Android, a fallback class was
implemented without using std::atomic, which is closer to the
previous implementation. Both classes appear to work properly after
performing a few dEQP tests. Both also perform similarly.

A few minor changes were made in order to attempt to reduce the use
of atomic integer operations when possible.

Change-Id: Ife6d3a2b6113346f8f8163b692e79c2a0e03b22f