Prevent initializing outline edges to out of bound values
When multisampling is enabled, outline edges were getting
initialized to one of the  primitive's X position. If the
primitive was out of bounds, then the default position was
out of bounds, which led to an initial out of bounds memory
access.

Added a clamp to fix the issue.

Bug chromium:779364

Change-Id: I4661f4229ee28a3032c763ed18dde799d3c3926b