Don't use macro expressions in another macro
We use icecc to speed up compilation in our project. After this change
https://swiftshader-review.googlesource.com/12308 chromium doesn't
compile via icecc as __has_include(<atomic>) is never expanded to 0 or 1
on icecc host and file <atomic> is not being included. My guess is that
icecc host gets partially expanded code and it's not being passed by
preprocessor again so condition #if __has_include(<atomic>) (expanded
from #if USE_STD_ATOMIC) is never fulfilled.
Bug: 783135
Change-Id: I09cf120c399a794a15dec05d6edd74c7848fc53a