Compute implicit derivatives relative to first quad pixel.
Previously the derivatives in y were computed using the difference
between the second and fourth pixel in the quad. For consistency with
the x derivative, use the first and third pixel instead. Also, some
shuffling can be eliminated by having the x derivative in the y
component, and the y derivative in the z component.

Change-Id: I985dcd3c5e2c47c10caf020cf5cb79587b3b3aab