Move draw call early-outs until after validation.
Drawing without a current program is not an error and we can early-out,
but not until after other validation which could generate an error has
completed. Validation happening on a valid program happens afterwards
though. Likewise, providing insufficient vertices for even one
primitive results in drawing nothing, but validation still needs to
happen.

Change-Id: I5385ffe352fc38343caa41eb99f5549472da3b4f